+++ CHANGELOG.md
@@ -25,6 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `apiRequireFaction` returns also user avatar
 ### Removed
 - `maybeFaction` function removed
+### Fixed
+- [Empty construction queue on a planet causes error during simulation][4]
 
 ## [0.1.0] - 2019-05-11
 ### Added
@@ -42,3 +44,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/tuturto/deep-sky/compare/0.1.0...HEAD
 [0.1.0]: https://github.com/tuturto/deep-sky/releases/tag/0.1.0
+[4]: https://github.com/tuturto/deep-sky/issues/4